requirements.txt
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
PyYAML==6.0.3
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
